+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * Closed [#720](https://github.com/appserver-io/appserver/issues/720) - Add Request::getProposedSessionId()
 * Closed [#715](https://github.com/appserver-io/appserver/issues/715) - Add setAttribute/getAttribute methods to Request
 * Closed [#714](https://github.com/appserver-io/appserver/issues/714) - Moved and refactored error, welcome and auto-index pages
+* Closed [#743](https://github.com/appserver-io/appserver/issues/743) - Move PhtmlServlet as DhtmlServlet from Routlt 2 (to avoid conflicts with simple PHTML templates)
+* Closed [#744](https://github.com/appserver-io/appserver/issues/744) - Add .dhtml file handler for ServletEngine
 
 # Version 1.1.0-alpha2